refactor(spaces): share colour maps, use the native menu components, merge duplicate specs
Ponytail audit, shared-spaces web + mobile UI section.

- Extract $lib/utils/space-colors.ts: the 10-entry UserAvatarColor gradient map
  was copy-pasted verbatim in 6 files and the bg/border/text families in 5 more.
  All values stay full literal class names so Tailwind's scanner still finds
  them. bg-primary (@immich/ui ramp, mode-aware) and bg-immich-primary (legacy
  token, mode-agnostic) are different colours in dark mode, so both are kept as
  separate fields. space-new-assets-divider keeps its own map on purpose - its
  shades genuinely differ and it has one consumer.
- Replace 4 hand-rolled dropdowns with the components the same folder already
  uses: ButtonContextMenu/MenuOption for the two kebab menus, and Dropdown for
  the three labelled sort/group triggers (ButtonContextMenu is icon-only and
  would drop the visible label - upstream AlbumsControls drives the identical
  menus with Dropdown). Playwright specs move to role+name queries accordingly.
- Merge the two space-people-page specs, which duplicated an ~80-line mock
  harness for the same route component; every distinct assertion is preserved.
- Mobile: share the search/sort bar and no-match widgets between the two grid
  pages; drop the onAlbumsPicked callback prop, which duplicated the page's
  pop-return channel that the only production caller already consumes (the test
  now drives a real push/pop instead); make the three SpaceAlbumsPage callbacks
  required rather than silently defaulting to no-ops.
- Drop the never-read storeIndex enum field, the computed-but-unrendered
  onboarding step descriptions, and the stale ownerId cast in
  space-album-grouping (the DTO has carried ownerId since it was added).

timelineOverviewControlsEnabled and syncBannerTopSliverHeight are NOT removed:
the audit read them as test-only, but the former is declared identically in 13
pages and the latter is one of a 3-member top-sliver-height contract asserted as
a set. Cutting one of each would be inconsistent.
