+++ web/src/lib/components/spaces/space-onboarding-banner.svelte
@@ -45,7 +45,6 @@
       id: 'add-photos',
       icon: mdiImagePlusOutline,
       label: $t('spaces_onboarding_add_photos'),
-      description: $t('spaces_onboarding_add_photos_description'),
       complete: hasPhotos,
       action: onAddPhotos,
     },
@@ -53,7 +52,6 @@
       id: 'invite-members',
       icon: mdiAccountPlusOutline,
       label: $t('spaces_onboarding_invite_members'),
-      description: $t('spaces_onboarding_invite_members_description'),
       complete: hasMembers,
       action: onInviteMembers,
     },
@@ -61,7 +59,6 @@
       id: 'set-cover',
       icon: mdiImageFilterHdrOutline,
       label: $t('spaces_onboarding_set_cover'),
-      description: $t('spaces_onboarding_set_cover_description'),
       complete: hasCover,
       action: onSetCover,
     },
