+++ web/src/lib/components/spaces/spaces-controls.svelte
@@ -1,9 +1,15 @@
 <script lang="ts">
+  import Dropdown from '$lib/elements/Dropdown.svelte';
   import { SortOrder } from '$lib/stores/preferences.store';
-  import { sortOptionsMetadata, SpaceSortBy, spaceViewSettings } from '$lib/stores/space-view.store';
+  import {
+    sortOptionsMetadata,
+    SpaceSortBy,
+    spaceViewSettings,
+    type SpaceSortOptionMetadata,
+  } from '$lib/stores/space-view.store';
   import type { SharedSpaceResponseDto } from '@immich/sdk';
   import { Icon } from '@immich/ui';
-  import { mdiArrowDownThin, mdiArrowUpThin, mdiFormatListBulletedSquare, mdiSort, mdiViewGridOutline } from '@mdi/js';
+  import { mdiArrowDownThin, mdiArrowUpThin, mdiFormatListBulletedSquare, mdiViewGridOutline } from '@mdi/js';
   import { t } from 'svelte-i18n';
 
   interface Props {
@@ -13,20 +19,17 @@
 
   let { spaces, onSorted }: Props = $props();
 
-  let showDropdown = $state(false);
-
   const flipOrdering = (ordering: string) => {
     return ordering === SortOrder.Asc ? SortOrder.Desc : SortOrder.Asc;
   };
 
-  const handleSort = (option: (typeof sortOptionsMetadata)[0]) => {
+  const handleSort = (option: SpaceSortOptionMetadata) => {
     if ($spaceViewSettings.sortBy === option.id) {
       $spaceViewSettings.sortOrder = flipOrdering($spaceViewSettings.sortOrder);
     } else {
       $spaceViewSettings.sortBy = option.id;
       $spaceViewSettings.sortOrder = option.defaultOrder;
     }
-    showDropdown = false;
   };
 
   const sortSpaces = (items: SharedSpaceResponseDto[], sortBy: string, sortOrder: string) => {
@@ -63,8 +66,10 @@
   });
 
   let sortIcon = $derived($spaceViewSettings.sortOrder === SortOrder.Desc ? mdiArrowDownThin : mdiArrowUpThin);
-  let activeLabel = $derived(
-    sortOptionsMetadata.find((o) => o.id === $spaceViewSettings.sortBy)?.label ?? ('last_activity' as const),
+  // Default is sort by last activity
+  const defaultSortOption = sortOptionsMetadata[1];
+  let selectedSortOption = $derived(
+    sortOptionsMetadata.find(({ id }) => id === $spaceViewSettings.sortBy) ?? defaultSortOption,
   );
 </script>
 
@@ -78,36 +83,11 @@
     <Icon icon={$spaceViewSettings.viewMode === 'card' ? mdiFormatListBulletedSquare : mdiViewGridOutline} size="18" />
   </button>
 
-  <button
-    type="button"
-    class="flex items-center gap-1 rounded-lg px-3 py-1.5 text-sm hover:bg-gray-100 dark:hover:bg-gray-800"
-    onclick={() => (showDropdown = !showDropdown)}
-    data-testid="sort-button"
-  >
-    <Icon icon={mdiSort} size="18" />
-    <span>{$t(activeLabel)}</span>
-    <Icon icon={sortIcon} size="18" />
-  </button>
-
-  {#if showDropdown}
-    <div
-      class="absolute top-full right-0 z-10 mt-1 min-w-[180px] rounded-lg border bg-white shadow-lg dark:border-gray-700 dark:bg-gray-900"
-      data-testid="sort-dropdown"
-    >
-      {#each sortOptionsMetadata as option (option.id)}
-        <button
-          type="button"
-          class="flex w-full items-center justify-between px-4 py-2 text-sm hover:bg-gray-100 dark:hover:bg-gray-800"
-          class:font-semibold={$spaceViewSettings.sortBy === option.id}
-          onclick={() => handleSort(option)}
-          data-testid="sort-option-{option.id}"
-        >
-          <span>{$t(option.label)}</span>
-          {#if $spaceViewSettings.sortBy === option.id}
-            <Icon icon={sortIcon} size="16" />
-          {/if}
-        </button>
-      {/each}
-    </div>
-  {/if}
+  <Dropdown
+    position="bottom-right"
+    options={sortOptionsMetadata}
+    selectedOption={selectedSortOption}
+    onSelect={handleSort}
+    render={({ label }) => ({ title: $t(label), icon: sortIcon })}
+  />
 </div>
