+++ web/src/routes/(user)/people/people-page.spec.ts
@@ -80,14 +80,12 @@ function makeSpacePerson(overrides: Partial<SharedSpacePersonResponseDto> = {}):
     id: 'space-person-1',
     spaceId: 'space-1',
     name: 'Shared Alice',
-    thumbnailPath: '',
     isHidden: false,
     birthDate: null,
     representativeFaceId: null,
     representativeFaceSource: RepresentativeFaceSource.Auto,
     faceCount: 1,
     assetCount: 4,
-    alias: null,
     createdAt: '2026-01-01T00:00:00.000Z',
     updatedAt: '2026-01-02T00:00:00.000Z',
     type: 'person',
