+++ web/src/routes/(user)/spaces/[spaceId]/[[photos=photos]]/[[assetId=id]]/spaces-page.spec.ts
@@ -97,11 +97,6 @@ vi.mock('$lib/components/filter-panel/search-sort-dropdown.svelte', async () => 
   return { default: MockComponent };
 });
 
-vi.mock('$lib/components/filter-panel/sort-toggle.svelte', async () => {
-  const { default: MockComponent } = await import('@test-data/mocks/sort-toggle.stub.svelte');
-  return { default: MockComponent };
-});
-
 vi.mock('$lib/components/search/smart-search-results.svelte', async () => {
   const { default: MockComponent } = await import('@test-data/mocks/smart-search-results.stub.svelte');
   return { default: MockComponent };
@@ -279,7 +274,6 @@ describe('Spaces page search URL state', () => {
     renderPage();
 
     expect(screen.queryByPlaceholderText(/search/i)).not.toBeInTheDocument();
-    expect(screen.queryByTestId('sort-toggle')).not.toBeInTheDocument();
     expect(screen.getByTestId('smart-search-results')).toHaveAttribute('data-search-query', 'beach');
     expect(screen.getByTestId('smart-search-results')).toHaveAttribute('data-sort-order', 'relevance');
   });
@@ -290,7 +284,6 @@ describe('Spaces page search URL state', () => {
     renderPage();
 
     expect(screen.queryByPlaceholderText(/search/i)).not.toBeInTheDocument();
-    expect(screen.queryByTestId('sort-toggle')).not.toBeInTheDocument();
     expect(screen.getByTestId('smart-search-results')).toHaveAttribute('data-search-query', 'beach');
     expect(screen.getByTestId('smart-search-results')).toHaveAttribute('data-sort-order', 'asc');
   });
