+++ web/src/routes/(user)/spaces/[spaceId]/people/page-load.spec.ts
@@ -45,14 +45,12 @@ describe('space people page load', () => {
       id: 'person-1',
       spaceId: 'space-1',
       name: 'Alice',
-      thumbnailPath: '',
       isHidden: false,
       birthDate: null,
       representativeFaceId: null,
       representativeFaceSource: 'auto',
       faceCount: 3,
       assetCount: 4,
-      alias: null,
       createdAt: '2026-01-01T00:00:00.000Z',
       updatedAt: '2026-01-02T00:00:00.000Z',
       type: 'person',
